refactor: move unit tests to src/unit-test/ and add E2E test docs
将所有单元测试文件从 src/ 根目录移至 src/unit-test/，减少源码目录混杂。
新增 src/unit-test/e2e/ 目录，按 PR 记录 E2E 手动测试步骤。

- 13 个 .test.ts 文件迁移，更新相对 import 路径
- 新增 pr-57-close-code-4001.md E2E 测试文档
